Establish separation between user facing and internal code
